requirements.txt
pytest == 8.4.1
pytest-playwright == 0.7.0
allure-pytest == 2.14.3
allure-python-commons==2.14.3
pytest-xdist == 3.7.0
requests == 2.32.4
pytest-sugar == 1.0.0
pytest-rerunfailures == 15.1
pytest-base-url == 2.1.0
loguru == 0.7.3
openpyxl == 3.1.5
